requirements.txt
Appium-Python-Client>=3.0.0
selenium>=4.15.0
